scalability parameters must be Integer

+++ scalability_explorer/src/test/java/eu/choreos/vv/increasefunctions/ScalabilityFunctionsTest.java
@@ -8,18 +8,18 @@ public class ScalabilityFunctionsTest {
 
 	@Test
 	public void shouldIncreaseLinearly() {
-		final double fixedParam = 2;
+		final int fixedParam = 2;
 		ScalabilityFunction function = new LinearIncrease(fixedParam);
-		for(double i = 1; i<=10; i+=fixedParam)
-		assertEquals(i+fixedParam, function.increaseParams(i));
+		for(int i = 1; i<=10; i+=fixedParam)
+		assertEquals(i+fixedParam, function.increaseParams(i).intValue());
 	}
 	
 	@Test
 	public void shouldIncreaseExponentialy() {
-		final double fixedParam = 2;
+		final int fixedParam = 2;
 		ScalabilityFunction function = new ExponentialIncrease(fixedParam);
-		for(double i = 1; i<=10; i*=fixedParam)
-		assertEquals(i*fixedParam, function.increaseParams(i));
+		for(int i = 1; i<=10; i*=fixedParam)
+		assertEquals(i*fixedParam, function.increaseParams(i).intValue());
 	}
 
 }
